prebuild: Android.mk: Add prebuilt-intermediates for xmlpool
The AOSP master build system has started to police the use of
non-hermetic tools from the system PATH. The Mesa project was using
xgettext and other tools which are not part of the standard Android host
build tools, and are therefore not allowed. This ban is now being
enforced.

To work around this problem, introduce prebuilt-intermediates for
xmlpool based on the current source tree.

Bug: 116125577
Change-Id: Ica53192b4c5495550ce5d6b49510b5984cedd4be
Signed-off-by: Alistair Strachan <astrachan@google.com>
[jstultz: Simplified to just the Android.mk changes]
Signed-off-by: John Stultz <john.stultz@linaro.org>