HACK: prebuild: Add intermediate prebuilt generation script
Try to add a script to generate/update the intermediate prebuilt files

This is really a short term hack. Long term we should try
to get a upstream target in meson so that we can generate
the intermediate source files using the normal build.

Then we can generate them before checking in a update to
AOSP.

Lots of credit to Amit Pundir for the original effort as
well as sorting out the xmlpool bits!

Change-Id: I6cf41fc627f5576b48397cee65821990c5fb3702
Signed-off-by: John Stultz <john.stultz@linaro.org>